+++ portal/lib/admin_pages/run_list_page_ui.dart
@@ -694,6 +694,7 @@ class RunListPage extends StatelessWidget {
                 allKennels: allKennels,
                 canViewMonitor: formController.canViewMonitor,
                 canManageNewsflash: formController.canManageNewsflash,
+                canManagePermissions: formController.canManagePermissions,
               ),
             );
           },